worker/pyproject.toml
[project]
name = "pan-worker"
version = "0.1.0"
description = "Worker for processing videos"
requires-python = ">=3.10"
dependencies = [
    "flask",
    "torch",
    "torchvision",
    "opencv-python",
    "ultralytics",
    "requests",
    "psutil",
    "pandas",
    "Pillow",
    "megadetector",
]
